checkpoint 1

+++ contracts/src/common/BridgeEncoder.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 import {BridgeMessageType} from "./IBridge.sol";
-import {TransferData, MigrationData} from "./TransferData.sol";
+import {TransferData} from "./TransferData.sol";
 
 /**
  * @dev Library for encoding and decoding bridge messages.
@@ -14,33 +14,34 @@ library BridgeEncoder {
     /// @dev Error thrown when message type is invalid for renewal
     error InvalidRenewalMessageType();
 
-    /**
-     * @dev Encode an ejection message.
-     */
+    /// @dev Encode an ejection message.
     function encodeEjection(
         bytes memory dnsEncodedName,
         TransferData memory data
     ) internal pure returns (bytes memory) {
-        return abi.encode(uint(BridgeMessageType.EJECTION), dnsEncodedName, data);
+        return
+            abi.encode(uint(BridgeMessageType.EJECTION), dnsEncodedName, data);
     }
 
-    /**
-     * @dev Decode an ejection message.
-     */
-    function decodeEjection(bytes memory message) internal pure returns (
-        bytes memory dnsEncodedName,
-        TransferData memory data
-    ) {
-        uint _messageType;
-        (_messageType, dnsEncodedName, data) = abi.decode(message, (uint, bytes, TransferData));
+    /// @dev Decode an ejection message.
+    function decodeEjection(
+        bytes memory message
+    )
+        internal
+        pure
+        returns (bytes memory dnsEncodedName, TransferData memory data)
+    {
+        uint256 _messageType;
+        (_messageType, dnsEncodedName, data) = abi.decode(
+            message,
+            (uint256, bytes, TransferData)
+        );
         if (_messageType != uint(BridgeMessageType.EJECTION)) {
             revert InvalidEjectionMessageType();
         }
     }
 
-    /**
-     * @dev Encode a renewal message.
-     */
+    /// @dev Encode a renewal message.
     function encodeRenewal(
         uint256 tokenId,
         uint64 newExpiry
@@ -48,25 +49,25 @@ library BridgeEncoder {
         return abi.encode(uint(BridgeMessageType.RENEWAL), tokenId, newExpiry);
     }
 
-    /**
-     * @dev Decode a renewal message.
-     */
-    function decodeRenewal(bytes memory message) internal pure returns (
-        uint256 tokenId,
-        uint64 newExpiry
-    ) {
-        uint _messageType;
-        (_messageType, tokenId, newExpiry) = abi.decode(message, (uint, uint256, uint64));
+    /// @dev Decode a renewal message.
+    function decodeRenewal(
+        bytes memory message
+    ) internal pure returns (uint256 tokenId, uint64 newExpiry) {
+        uint256 _messageType;
+        (_messageType, tokenId, newExpiry) = abi.decode(
+            message,
+            (uint256, uint256, uint64)
+        );
         if (_messageType != uint(BridgeMessageType.RENEWAL)) {
             revert InvalidRenewalMessageType();
         }
     }
 
-    /**
-     * @dev Helper function to get the message type from an encoded message.
-     */
-    function getMessageType(bytes memory message) internal pure returns (BridgeMessageType) {
-        uint _messageType = abi.decode(message, (uint));
+    /// @dev Helper function to get the message type from an encoded message.
+    function getMessageType(
+        bytes memory message
+    ) internal pure returns (BridgeMessageType) {
+        uint256 _messageType = abi.decode(message, (uint256));
         return BridgeMessageType(_messageType);
     }
 }
